Refactor StepByStepFilter class
This allows us to expand the StepByStepFilter service without
overloading the class with unrelated methods. It will be useful in the
next commit when we add the functionality to get the statuses that can
be filtered by from a single source.
